# apps/api/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: "Afrik'Fid API"
  version: "1.0.0"
  description: |
    Passerelle de paiement B2B multi-pays avec fidélisation intégrée pour l'Afrique francophone et anglophone.

    ## Authentification

    L'API supporte deux mécanismes d'authentification :

    **1. JWT Bearer** — Pour le tableau de bord admin et marchand
    ```
    Authorization: Bearer <accessToken>
    ```
    Obtenir un token via `POST /auth/admin/login` ou `POST /auth/merchant/login`.

    **2. API Key** — Pour les intégrations marchand (paiements, confirmations, statuts)
    ```
    X-API-Key: af_pub_xxxxxxxxxxxx
    ```
    Ajouter `X-Sandbox: true` pour utiliser les clés sandbox et simuler les paiements.

    ## Modèle de fidélité X/Y/Z

    Chaque paiement est réparti entre trois parties :
    - **X%** — Commission marchand (remise accordée au client ou gardée)
    - **Y%** — Rebate client (selon statut loyalty : OPEN=0%, LIVE=5%, GOLD=8%, ROYAL=10%)
    - **Z%** — Commission plateforme Afrik'Fid

    ## Environnements

    | Environnement | Base URL | Clé |
    |---|---|---|
    | Production | `https://api.afrikfid.com/api/v1` | `af_pub_xxx` |
    | Sandbox | `https://api.afrikfid.com/api/v1` + `X-Sandbox: true` | `af_sandbox_pub_xxx` |
    | Local | `http://localhost:4001/api/v1` | Clé de test |

  contact:
    name: "Support Afrik'Fid"
    email: support@afrikfid.com
  license:
    name: Propriétaire
servers:
  - url: http://localhost:4001/api/v1
...
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /auth/admin/login ou /auth/merchant/login
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API publique du marchand (af_pub_xxx en prod, af_sandbox_pub_xxx en sandbox)
    HmacSignature:
      type: apiKey
      in: header
      name: X-AfrikFid-Signature
      description: |
        Signature HMAC-SHA256 optionnelle pour l'intégrité des requêtes (CDC §4.5.2).
        Format : `sha256=<hex>`
        Calcul : `HMAC-SHA256(api_key_secret, raw_request_body)`
        En mode `REQUIRE_HMAC_SIGNATURE=true`, ce header devient obligatoire sur `/payments/initiate` et `/payments/{id}/refund`.

  # ─── Schemas ───────────────────────────────────────────────────────────────
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: "Données invalides"
      required: [error]
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
...
    Transaction:
      type: object
      properties:
        id: { type: string, format: uuid }
        reference: { type: string, example: "AFD-1710000000-ABC123" }
        merchantId: { type: string, format: uuid }
        clientId: { type: string, format: uuid, nullable: true }
        grossAmount: { type: number, example: 10000 }
        currency: { type: string, example: XOF }
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        paymentMethod:
          type: string
          enum: [MOBILE_MONEY, card]
        paymentOperator:
          type: string
          example: ORANGE
          nullable: true
        operatorRef: { type: string, nullable: true }
        clientLoyaltyStatus:
          type: string
          enum: [OPEN, LIVE, GOLD, ROYAL]
        merchantRebatePercent: { type: number }
        clientRebatePercent: { type: number }
        platformCommissionPercent: { type: number }
        merchantRebateAmount: { type: number }
        clientRebateAmount: { type: number }
        platformCommissionAmount: { type: number }
        merchantReceives: { type: number }
        rebateMode: { type: string, enum: [cashback, immediate] }
        initiatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
        expiresAt: { type: string, format: date-time }
    Merchant:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: "Boutique Koffi" }
        email: { type: string, format: email }
        phone: { type: string, example: "+2250701000001" }
        countryId: { type: string, example: CI }
        countryName: { type: string, example: "Côte d'Ivoire" }
        currency: { type: string, example: XOF }
        category: { type: string, example: retail }
        rebatePercent: { type: number, example: 8 }
        rebateMode: { type: string, enum: [cashback, immediate] }
        settlementFrequency: { type: string, enum: [daily, weekly, monthly] }
        webhookUrl: { type: string, format: uri, nullable: true }
        status: { type: string, enum: [pending, active, suspended] }
        kycStatus: { type: string, enum: [pending, approved, rejected] }
        isActive: { type: integer, enum: [0, 1] }
        apiKeyPublic: { type: string, example: "af_pub_abc123" }
        sandboxKeyPublic: { type: string, example: "af_sandbox_pub_abc123" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: Token manquant ou invalide
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: "TOKEN_INVALID", message: "Token invalide" }
...
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Données invalides
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }

# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /merchants:
    post:
      tags: [Merchants]
      summary: Créer un marchand (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, country_id, rebate_percent, rebate_mode]
              properties:
                name: { type: string, example: "Boutique Koffi" }
                email: { type: string, format: email }
                phone: { type: string }
                country_id: { type: string, example: CI }
                category: { type: string, example: retail }
                rebate_percent:
                  type: number
                  minimum: 0
                  maximum: 50
                  example: 8
                rebate_mode: { type: string, enum: [cashback, immediate] }
                password: { type: string, description: "Mot de passe dashboard marchand" }
                webhook_url: { type: string, format: uri }
                settlement_frequency: { type: string, enum: [daily, weekly, monthly] }
      responses:
        '201':
          description: Marchand créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant: { $ref: '#/components/schemas/Merchant' }
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Email déjà utilisé

    get:
      tags: [Merchants]
      summary: Lister les marchands (admin, paginé)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: status
          schema: { type: string, enum: [pending, active, suspended] }
        - in: query
          name: country_id
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
          description: Recherche par nom ou email
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchants:
                    type: array
                    items: { $ref: '#/components/schemas/Merchant' }
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
  /merchants/register:
    post:
      tags: [Merchants]
      summary: Inscription self-service (demande en attente)
      description: Crée un compte marchand avec statut `pending`. L'équipe Afrik'Fid valide sous 24-48h.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, phone, country_id, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                phone: { type: string }
                country_id: { type: string }
                password: { type: string, minLength: 8 }
                category: { type: string }
                rebate_percent: { type: number }
                rebate_mode: { type: string }
                webhook_url: { type: string, format: uri }
                website: { type: string, format: uri }
      responses:
        '201':
          description: Demande enregistrée
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  id: { type: string }
        '409':
          description: Email déjà utilisé
  /merchants/{id}:
    get:
      tags: [Merchants]
      summary: Détails d'un marchand (admin)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant: { $ref: '#/components/schemas/Merchant' }
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Merchants]
      summary: Mettre à jour un marchand (admin)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [pending, active, suspended] }
                kyc_status: { type: string, enum: [pending, approved, rejected] }
                rebate_percent: { type: number }
                rebate_mode: { type: string }
                webhook_url: { type: string, format: uri }
                settlement_frequency: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant: { $ref: '#/components/schemas/Merchant' }
  /merchants/me/profile:
    get:
      tags: [Merchants]
      summary: Profil du marchand connecté
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant: { $ref: '#/components/schemas/Merchant' }
  /merchants/me/stats:
    get:
      tags: [Merchants]
      summary: Statistiques du marchand connecté
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  stats:
                    type: object
                    properties:
                      total_transactions: { type: integer }
                      total_volume: { type: number }
                      total_received: { type: number }
                      total_rebate_given: { type: number }
                      completed_count: { type: integer }
                      failed_count: { type: integer }
                      pending_count: { type: integer }
                  byLoyaltyStatus:
                    type: array
                    items:
                      type: object
                      properties:
                        client_loyalty_status: { type: string }
                        count: { type: integer }
                        volume: { type: number }
  /merchants/me/transactions:
    get:
      tags: [Merchants]
      summary: Transactions du marchand connecté
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items: { $ref: '#/components/schemas/Transaction' }
                  total: { type: integer }

  # ── CLIENTS ───────────────────────────────────────────────────────────────
...